requirements.txt
python-dotenv
langchain
langgraph
langchain-community
langchain-openai
typing-extensions
openai
fastapi[standard]
tiktoken